Refactoring lrama/states.rb
* Separated State class from lrama/states.rb file
* Separated Lrama::State::Reduce class and Lrama::State::Shift class into separate files
