+++ lib/lrama.rb
@@ -7,6 +7,7 @@ require "lrama/lexer"
 require "lrama/output"
 require "lrama/parser"
 require "lrama/report"
+require "lrama/state"
 require "lrama/states"
 require "lrama/states_reporter"
 require "lrama/version"
